attach LOA pdf document to email
